[C] Warnings and formatting tidy up in client.

+++ aeron-client/src/main/c/aeron_subscription.c
@@ -70,6 +70,7 @@ int aeron_subscription_create(
     _subscription->is_closed = false;
 
     *subscription = _subscription;
+
     return 0;
 }
 
@@ -87,6 +88,7 @@ int aeron_subscription_delete(aeron_subscription_t *subscription)
 
     aeron_free((void *)subscription->channel);
     aeron_free(subscription);
+
     return 0;
 }
 
@@ -136,6 +138,7 @@ int aeron_subscription_alloc_image_list(volatile aeron_image_list_t **image_list
     _image_list->next_list = NULL;
 
     *image_list = _image_list;
+
     return 0;
 }
 
@@ -201,6 +204,7 @@ int aeron_client_conductor_subscription_install_new_image_list(
     image_list->next_list = subscription->conductor_fields.image_lists_head.next_list;
 
     AERON_PUT_ORDERED(subscription->conductor_fields.image_lists_head.next_list, image_list);
+
     return 0;
 }
 
@@ -270,6 +274,7 @@ int aeron_subscription_constants(aeron_subscription_t *subscription, aeron_subsc
     constants->on_available_image = subscription->on_available_image;
     constants->on_unavailable_image = subscription->on_unavailable_image;
     constants->channel_status_indicator_id = subscription->channel_status_indicator_id;
+
     return 0;
 }
 
@@ -282,8 +287,7 @@ int aeron_subscription_image_count(aeron_subscription_t *subscription)
     return image_list->length;
 }
 
-aeron_image_t *aeron_subscription_image_by_session_id(
-    aeron_subscription_t *subscription, int32_t session_id)
+aeron_image_t *aeron_subscription_image_by_session_id(aeron_subscription_t *subscription, int32_t session_id)
 {
     volatile aeron_image_list_t *image_list;
     aeron_image_t *result = NULL;
@@ -363,6 +367,7 @@ int aeron_subscription_image_retain(aeron_subscription_t *subscription, aeron_im
         subscription, aeron_subscription_last_image_list_change_number(subscription));
 
     aeron_image_incr_refcnt(image);
+
     return 0;
 }
 
@@ -383,6 +388,7 @@ int aeron_subscription_image_release(aeron_subscription_t *subscription, aeron_i
         subscription, aeron_subscription_last_image_list_change_number(subscription));
 
     aeron_image_decr_refcnt(image);
+
     return 0;
 }
 
@@ -446,7 +452,8 @@ int aeron_subscription_poll(
 int aeron_subscription_controlled_poll(
     aeron_subscription_t *subscription,
     aeron_controlled_fragment_handler_t handler,
-    void *clientd, size_t fragment_limit)
+    void *clientd,
+    size_t fragment_limit)
 {
     volatile aeron_image_list_t *image_list;
 
@@ -537,13 +544,15 @@ int aeron_header_values(aeron_header_t *header, aeron_header_values_t *values)
 
     memcpy(&values->frame, header->frame, sizeof(aeron_header_values_frame_t));
     values->initial_term_id = header->initial_term_id;
+
     return 0;
 }
 
 int64_t aeron_header_position(aeron_header_t *header)
 {
-    const int64_t offset_at_end_of_frame = AERON_ALIGN(
+    const int32_t offset_at_end_of_frame = AERON_ALIGN(
         header->frame->term_offset + header->frame->frame_header.frame_length, AERON_LOGBUFFER_FRAME_ALIGNMENT);
+
     return aeron_logbuffer_compute_position(
         header->frame->term_id, offset_at_end_of_frame, header->position_bits_to_shift, header->initial_term_id);
 }
